+++ Lib9c.Models.Tests/Stats/StatMapTest.cs
@@ -1,5 +1,5 @@
 using Bencodex.Types;
-using Lib9c.Models.Stat;
+using Lib9c.Models.Stats;
 
 namespace Lib9c.Models.Tests.Stats;
