+++ Lib9c.Models/AttachmentActionResults/ItemEnhancement11Result.cs
@@ -2,7 +2,7 @@ using System.Numerics;
 using Bencodex.Types;
 using Lib9c.Models.Exceptions;
 using Lib9c.Models.Factories;
-using Lib9c.Models.Item;
+using Lib9c.Models.Items;
 using Libplanet.Types.Assets;
 using Nekoyume.Action;
 using Nekoyume.Model.State;
