+++ Lib9c.Models/AttachmentActionResults/ItemEnhancement9Result.cs
@@ -2,7 +2,7 @@ using System.Numerics;
 using Bencodex.Types;
 using Lib9c.Models.Exceptions;
 using Lib9c.Models.Factories;
-using Lib9c.Models.Item;
+using Lib9c.Models.Items;
 using Nekoyume.Action;
 using Nekoyume.Model.State;
 using ValueKind = Bencodex.Types.ValueKind;
