+++ Lib9c.Models/AttachmentActionResults/RapidCombination5Result.cs
@@ -1,7 +1,7 @@
 using Bencodex.Types;
 using Lib9c.Models.Exceptions;
 using Lib9c.Models.Factories;
-using Lib9c.Models.Item;
+using Lib9c.Models.Items;
 using Nekoyume.Model.State;
 using ValueKind = Bencodex.Types.ValueKind;
