+++ Lib9c.Models/Items/Inventory.cs
@@ -3,7 +3,7 @@ using Bencodex.Types;
 using Lib9c.Models.Exceptions;
 using ValueKind = Bencodex.Types.ValueKind;
 
-namespace Lib9c.Models.Item;
+namespace Lib9c.Models.Items;
 
 /// <summary>
 /// <see cref="Nekoyume.Model.Item.Inventory"/>
