+++ Lib9c.Models/Items/ItemUsable.cs
@@ -3,11 +3,11 @@ using Bencodex.Types;
 using Lib9c.Models.Exceptions;
 using Lib9c.Models.Factories;
 using Lib9c.Models.Skills;
-using Lib9c.Models.Stat;
+using Lib9c.Models.Stats;
 using Nekoyume.Model.State;
 using ValueKind = Bencodex.Types.ValueKind;
 
-namespace Lib9c.Models.Item;
+namespace Lib9c.Models.Items;
 
 /// <summary>
 /// <see cref="Nekoyume.Model.Item.ItemUsable"/>
