+++ Lib9c.Models/Items/Material.cs
@@ -6,7 +6,7 @@ using Libplanet.Common;
 using Nekoyume.Model.State;
 using ValueKind = Bencodex.Types.ValueKind;
 
-namespace Lib9c.Models.Item;
+namespace Lib9c.Models.Items;
 
 /// <summary>
 /// <see cref="Nekoyume.Model.Item.Material"/>
