+++ Lib9c.Models/Items/Weapon.cs
@@ -1,6 +1,6 @@
 using Bencodex.Types;
 
-namespace Lib9c.Models.Item;
+namespace Lib9c.Models.Items;
 
 /// <summary>
 /// <see cref="Nekoyume.Model.Item.Weapon"/>
