+++ Lib9c.Models/Market/ItemProduct.cs
@@ -1,7 +1,7 @@
 using Bencodex;
 using Bencodex.Types;
 using Lib9c.Models.Factories;
-using Lib9c.Models.Item;
+using Lib9c.Models.Items;
 using Nekoyume.Model.State;
 
 namespace Lib9c.Models.Market;
