+++ Mimir/GraphQL/Types/EquipmentType.cs
@@ -1,5 +1,5 @@
 using Lib9c.GraphQL.Types;
-using Lib9c.Models.Item;
+using Lib9c.Models.Items;
 
 namespace Mimir.GraphQL.Types;
