+++ Mimir/Repositories/RuneSlotRepository.cs
@@ -1,5 +1,5 @@
 using Libplanet.Crypto;
-using Lib9c.Models.Rune;
+using Lib9c.Models.Runes;
 using Mimir.Enums;
 using Mimir.Exceptions;
 using Mimir.Services;
